Update 4-authentication-and-permissions.md
TypeError: Got a `TypeError` when calling `Snippet.objects.create()`. This may be because you have a writable 
field on the serializer class that is not a valid argument to `Snippet.objects.create()`. You may need to make
 the field read-only, or override the SnippetSerializer.create() method to handle this correctly.
Original exception text was: int() argument must be a string or a number, not 'SimpleLazyObject'.